+++ bot_modules/experimental/!README.txt
@@ -1,0 +1,1 @@
+Not actually used, previous experiments, mostly failed, but still worth preserving :-)
