+++ bot_modules/external/!readme.txt
@@ -1,0 +1,2 @@
+External modules, stuff that communicates with something outside of the bot.
+For example communication with Heroku, or external commands that are not part of the original beq design.
